Moved content-company.js to common
