+++ packages/common/graphql/fragments/content-company.js
