+++ sites/packworld/server/routes/content.js
@@ -3,7 +3,7 @@ const content = require('../templates/content');
 const company = require('../templates/content/company');
 const webinar = require('../templates/content/webinar');
 const queryFragment = require('../graphql/fragments/content-page');
-const companyQueryFragment = require('../graphql/fragments/content-company');
+const companyQueryFragment = require('../../../../packages/common/graphql/fragments/content-company');
 
 module.exports = (app) => {
   app.get('/*?company/:id(\\d{8})*', withContent({
